+ Backup Simple Works! (tested on IAP Chassis) + Closest NinetyDegree works as well (tested on IAP)

+++ 2024-Competition-Code/src/main/java/frc/robot/subsystems/swerve/SwerveDrive.java
@@ -276,7 +276,7 @@ public class SwerveDrive extends SubsystemBase {
       return field;
    }
 
-   public boolean inThreshold(int degrees){
+   public boolean inThreshold(double degrees){
       //Used in NinetyDegreeRotation.java\
       return (-degrees <= -this.getRotation().getDegrees() % 90) && (-this.getRotation().getDegrees() % 90 <= degrees);
    }
